+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Allow rendering of non-highlighted `ItemAction`.
 - [...]
 
 # v16.0.3 (09/12/2019)